fix(context-menu): address remaining review follow-ups
